--- service/data.csv
@@ -1,5 +1,0 @@
-x1,x2,x3,s1,s3,a2,a3,p,condition
-1,1,0,1,0,0.0,0.0,0,20
-1,0,1,0,0,1.0,0.0,0,5
-0,1,1,0,-1,0.0,1.0,0,10
--1,1,-3,0,0,1000000000,1000000000,1,0
